pyproject.toml
[tool.poetry]
name = "imageresizerwater"
version = "0.1.0"
description = "Mass image processing with resizing and watermarking"
authors = ["enrico"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.3.2"
progress = "^1.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
